# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kubeasy CLI API",
    "version": "1.0.0",
    "description": "Public API for the Kubeasy platform consumed by the CLI. Web routes use session cookie authentication; CLI routes use Bearer token authentication."
  },
  "servers": [
    {
      "url": "https://kubeasy.dev",
      "description": "Production"
    },
    {
      "url": "http://localhost:3024",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie set after login via better-auth"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained via `kubeasy login`. Used by the CLI and accepted on all public API routes."
      }
    },
...
  },
  "paths": {
...
    "/api/challenges": {
      "get": {
        "operationId": "listChallenges",
        "summary": "List challenges",
        "tags": ["CLI"],
        "security": [
          {
            "SessionAuth": []
          },
          {
            "BearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["easy", "medium", "hard"]
            },
            "required": false,
            "name": "difficulty",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "theme",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "showCompleted",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "theme": {
                            "type": "string"
                          },
                          "themeSlug": {
                            "type": "string"
                          },
                          "difficulty": {
                            "type": "string",
                            "enum": ["easy", "medium", "hard"]
                          },
                          "type": {
                            "type": "string"
                          },
                          "typeSlug": {
                            "type": "string"
                          },
                          "estimatedTime": {
                            "type": "integer"
                          },
                          "initialSituation": {
                            "type": "string"
                          },
                          "ofTheWeek": {
                            "type": "boolean"
                          },
                          "completedCount": {
                            "type": "integer"
                          },
                          "userStatus": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "slug",
                          "title",
                          "description",
                          "theme",
                          "themeSlug",
                          "difficulty",
                          "type",
                          "typeSlug",
                          "estimatedTime",
                          "initialSituation",
                          "ofTheWeek",
                          "completedCount",
                          "userStatus"
                        ]
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  },
                  "required": ["challenges", "count"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cli/track/login": {
      "post": {
        "operationId": "trackCliLogin",
        "summary": "Track CLI login and metadata",
        "tags": ["CLI"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliVersion": {
                    "type": "string"
                  },
                  "os": {
                    "type": "string"
                  },
                  "arch": {
                    "type": "string"
                  }
                },
                "required": ["cliVersion", "os", "arch"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login tracked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstLogin": {
                      "type": "boolean"
                    }
                  },
                  "required": ["firstLogin"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/cli/track/setup": {
      "post": {
        "operationId": "trackSetup",
        "summary": "Track CLI cluster initialisation",
        "tags": ["CLI"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliVersion": {
                    "type": "string"
                  },
                  "os": {
                    "type": "string"
                  },
                  "arch": {
                    "type": "string"
                  }
                },
                "required": ["cliVersion", "os", "arch"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setup tracked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "firstTime": {
                      "type": "boolean"
                    }
                  },
                  "required": ["success", "firstTime"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    }
  }
}